+++ drf_yasg_json_api/deprecation.py
@@ -1,0 +1,2 @@
+class DrfYasgJsonApiDeprecationWarning(DeprecationWarning):
+    pass