+++ tests/base.py
@@ -1,4 +1,4 @@
-import drf_yasg
+import drf_yasg.inspectors
 
 import drf_yasg_json_api.inspectors
 
@@ -6,7 +6,8 @@ import drf_yasg_json_api.inspectors
 class BasicSwaggerAutoSchema(drf_yasg_json_api.inspectors.SwaggerAutoSchema):
     field_inspectors = [
         drf_yasg_json_api.inspectors.NamesFormatFilter,
-        drf_yasg_json_api.inspectors.InlineSerializerSmartInspector,
+        drf_yasg_json_api.inspectors.JSONAPISerializerSmartInspector,
+        drf_yasg.inspectors.ReferencingSerializerInspector,
         drf_yasg_json_api.inspectors.IntegerIDFieldInspector,
         drf_yasg_json_api.inspectors.IntegerPrimaryKeyRelatedFieldInspector,
         drf_yasg_json_api.inspectors.ManyRelatedFieldInspector,